Add files via upload
In this we have to Create an AVL tree application that uses a tree structure containing all of the words in a document , with a count of the number of times each word is used. Each entry in the AVL tree contains a word from the document and a pointer to an integer that contains a count of the number of times the word appears in the document.